requirements.txt
discord.py
trello.py